refactor ssh config into the ssh file
